+++ tools/rosetta/converter_test.go
@@ -1,25 +1,22 @@
 package rosetta_test
 
 import (
-	"cosmossdk.io/tools/rosetta"
 	"encoding/hex"
 	"encoding/json"
 	"testing"
 
-	abci "github.com/tendermint/tendermint/abci/types"
+	"cosmossdk.io/tools/rosetta"
+	crgerrs "cosmossdk.io/tools/rosetta/lib/errors"
 
-	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
+	rosettatypes "github.com/coinbase/rosetta-sdk-go/types"
+	"github.com/stretchr/testify/suite"
+	abci "github.com/tendermint/tendermint/abci/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-
-	rosettatypes "github.com/cosmos/rosetta-sdk-go/types"
-	"github.com/stretchr/testify/suite"
-
-	crgerrs "cosmossdk.io/tools/rosetta/lib/errors"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	bank "github.com/cosmos/cosmos-sdk/x/bank/types"
 )