+++ x/group/internal/orm/table_test.go
@@ -4,14 +4,16 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/cosmos/gogoproto/proto"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/group/errors"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestNewTable(t *testing.T) {
@@ -20,7 +22,7 @@ func TestNewTable(t *testing.T) {
 
 	testCases := []struct {
 		name        string
-		model       codec.ProtoMarshaler
+		model       proto.Message
 		expectErr   bool
 		expectedErr string
 	}{
@@ -54,7 +56,7 @@ func TestNewTable(t *testing.T) {
 func TestCreate(t *testing.T) {
 	specs := map[string]struct {
 		rowID  RowID
-		src    codec.ProtoMarshaler
+		src    proto.Message
 		expErr *sdkerrors.Error
 	}{
 		"empty rowID": {
@@ -122,7 +124,7 @@ func TestCreate(t *testing.T) {
 
 func TestUpdate(t *testing.T) {
 	specs := map[string]struct {
-		src    codec.ProtoMarshaler
+		src    proto.Message
 		expErr *sdkerrors.Error
 	}{
 		"happy path": {