+++ x/group/internal/orm/types.go
@@ -7,6 +7,8 @@ import (
 	"io"
 	"reflect"
 
+	"github.com/cosmos/gogoproto/proto"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -79,7 +81,7 @@ type Iterator interface {
 	// LoadNext loads the next value in the sequence into the pointer passed as dest and returns the key. If there
 	// are no more items the ErrORMIteratorDone error is returned
 	// The key is the rowID.
-	LoadNext(dest codec.ProtoMarshaler) (RowID, error)
+	LoadNext(dest proto.Message) (RowID, error)
 	// Close releases the iterator and should be called at the end of iteration
 	io.Closer
 }
@@ -93,18 +95,18 @@ type Indexable interface {
 }
 
 // AfterSetInterceptor defines a callback function to be called on Create + Update.
-type AfterSetInterceptor func(store sdk.KVStore, rowID RowID, newValue, oldValue codec.ProtoMarshaler) error
+type AfterSetInterceptor func(store sdk.KVStore, rowID RowID, newValue, oldValue proto.Message) error
 
 // AfterDeleteInterceptor defines a callback function to be called on Delete operations.
-type AfterDeleteInterceptor func(store sdk.KVStore, rowID RowID, value codec.ProtoMarshaler) error
+type AfterDeleteInterceptor func(store sdk.KVStore, rowID RowID, value proto.Message) error
 
 // RowGetter loads a persistent object by row ID into the destination object. The dest parameter must therefore be a pointer.
 // Any implementation must return `sdkerrors.ErrNotFound` when no object for the rowID exists
-type RowGetter func(store sdk.KVStore, rowID RowID, dest codec.ProtoMarshaler) error
+type RowGetter func(store sdk.KVStore, rowID RowID, dest proto.Message) error
 
 // NewTypeSafeRowGetter returns a `RowGetter` with type check on the dest parameter.
 func NewTypeSafeRowGetter(prefixKey [2]byte, model reflect.Type, cdc codec.Codec) RowGetter {
-	return func(store sdk.KVStore, rowID RowID, dest codec.ProtoMarshaler) error {
+	return func(store sdk.KVStore, rowID RowID, dest proto.Message) error {
 		if len(rowID) == 0 {
 			return sdkerrors.Wrap(errors.ErrORMEmptyKey, "key must not be nil")
 		}
@@ -122,7 +124,7 @@ func NewTypeSafeRowGetter(prefixKey [2]byte, model reflect.Type, cdc codec.Codec
 	}
 }
 
-func assertCorrectType(model reflect.Type, obj codec.ProtoMarshaler) error {
+func assertCorrectType(model reflect.Type, obj proto.Message) error {
 	tp := reflect.TypeOf(obj)
 	if tp.Kind() != reflect.Ptr {
 		return sdkerrors.Wrap(sdkerrors.ErrInvalidType, "model destination must be a pointer")